Добавляет данные из master для корректного слияния
